Bringing the two projects together
The complexity and overhead of managing two projects was pretty significant, and the demo project itself was getting complicated enough that I didn't think it was achieving its stated goals. Having a single project will simplify my dev work as well as deployment concerns for anyone building their own MUDs. It was a good experiment and I'm glad I tried it, but it's time to move forward.