Extract CreatureFactory service
Pulled out the logic to find an existing player's Creature or create a new one into a service. Right now it's quite simple but as Creature gets more and more options this Factory is going to become very important. It will be the place where you trade in the template for your PC or NPC and get back a fully built and ready-to-roam-around-the-world Creature.